restructure: io/logging -> utils/logging
