+++ io/cloudinary/fetchCloudinaryImageMetadata.ts
@@ -1,7 +1,7 @@
 import { filesystemCache, type CacheAdapter } from '@/io/cache/adapter'
 import cloudinary, { type CloudinaryClient } from '@/io/cloudinary/client'
-import { getErrorDetails } from '@/io/logging/logging'
-import { formatValidationError } from '@/io/logging/zod'
+import { getErrorDetails } from '@/utils/logging/logging'
+import { formatValidationError } from '@/utils/logging/zod'
 import parsePublicIdFromCloudinaryUrl from './parsePublicIdFromCloudinaryUrl'
 import { Ok, toErr, type Result } from '@/utils/errors/result'
 import { z } from 'zod'
