+++ io/notion/getPost.ts
@@ -4,7 +4,7 @@ import getBlockChildren from '@/io/notion/getBlockChildren'
 import getPosts from '@/io/notion/getPosts'
 import { PostListItemSchema, PostPropertiesSchema, type Post } from './schemas/post'
 import { PostPageMetadataSchema } from './schemas/page'
-import { logValidationError } from '@/io/logging/zod'
+import { logValidationError } from '@/utils/logging/zod'
 import { env } from '@/io/env/env'
 import { Ok, Err, toErr, type Result } from '@/utils/errors/result'
 import { withRetry } from '@/io/retry'
