+++ io/notion/getPosts.ts
@@ -2,7 +2,7 @@ import { filesystemCache, type CacheAdapter } from '@/io/cache/adapter'
 import notion, { collectPaginatedAPI, type Client } from './client'
 import { PostListItemSchema, PostPropertiesSchema, type PostListItem } from './schemas/post'
 import { PageMetadataSchema } from './schemas/page'
-import { logValidationError } from '@/io/logging/zod'
+import { logValidationError } from '@/utils/logging/zod'
 import { env } from '@/io/env/env'
 import { type Result, Ok, toErr } from '@/utils/errors/result'
 import { withRetry } from '@/io/retry'
